Finished up the tutoring services page
I added proper table elements to the page, as well as fixed the list items under subjects, and fixed the size of each section so they arent super big or super small.  The base website is finished now.  All thats left to do is make changes to the logo, maybe change the two main colors to something else, and little changes here and there to make the site look prettier.  I also need to find a way to host the site publicly so others can visit it.