Made huge changes to the home page in the CSS and HTML
I had to redo a lot of the CSS mostly (and some HTML) of the home page because originally when viewed on a mobile screen, there was a BIG white space on the right side of the screen due to smaller size of phone.  I went and changed things like the padding/margin of several elements in the front page (the header, the youtube videos, and paragraphs of text).  After doing this, the home page was no longer stuck scrolling when viewed in a smaller window and therefore it should be okay on phone screens now.